fix: remove duplication of fixtures
