+++ tests/test_models/test_prophet.py
@@ -16,18 +16,6 @@ from tests.test_models.utils import assert_prediction_components_are_present
 from tests.test_models.utils import assert_sampling_is_valid
 
 
-@pytest.fixture
-def ts_with_external_timestamp():
-    df = generate_ar_df(periods=100, start_time=10, n_segments=2, freq=None)
-    df_wide = TSDataset.to_dataset(df)
-    df_exog = generate_ar_df(periods=100, start_time=10, n_segments=2, freq=None)
-    df_exog["target"] = pd.date_range(start="2020-01-01", periods=100).tolist() * 2
-    df_exog_wide = TSDataset.to_dataset(df_exog)
-    df_exog_wide.rename(columns={"target": "external_timestamp"}, level="feature", inplace=True)
-    ts = TSDataset(df=df_wide.iloc[:-10], df_exog=df_exog_wide, known_future="all", freq=None)
-    return ts
-
-
 def _check_forecast(ts, model, horizon):
     model.fit(ts)
     future_ts = ts.make_future(future_steps=horizon)
